Added Techboxes, some structure changes.
Created Techboxes (the "part" portion of #5, #6, #7 and #8)
Resolved #8 - SCU logic was pretty easy.
Created general framework to allow beacons to check and report which
tech modules are attached and active.
Split out ESLDHailer and ESLDBeacon into their own files.
Modified code to address #16, but I still need to test that it's
working.