Worked on Techboxes, gravity, spread.
Worked on #5, crew death and null cases are working but there's a sticky
enumeration issue on part killing when it's full of explodables.
Resolved #6 - it may need a balance pass, but it seems to be working
fine and relative velocities are coming through OK.
Worked on #7 - Functionality is there, but it seems to be bad at
recognizing when it gets turned off.
Resolved #14, or at least enough that I'm willing to call it good for
now until I've got some input on balancing.
Resolved #16 - I couldn't make it put me too far away from the arrival
spot, so I think I fixed it.