Add ESLDSampleTank class
This addresses #39 - need to texture the part before I close it, but
here's the functionality.  The module wipes out most of the canister's
value (to an extent modifiable in the cfg) when any resources are
removed from the tank, preventing abuse of vehicle recovery to get
around forced scarcity.  @BobPalmer, you may be interested.